chore(library): Splitter library in multiple files
