+++ .npmignore
@@ -1,12 +1,9 @@
-tsconfig.json
+.github
+src
 test
+.gitignore
+.prettierrc
 babel.config.js
 jest.config.js
-index.js.map
-shared/index.js.map
-vanilla.js.map
-.prettierrc
-src
-.gitignore
+tsconfig.json
 yarn.lock
-.github
