+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "http-react",
-  "version": "2.4.7",
+  "version": "2.4.8",
   "description": "React hooks for data fetching",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
     "test": "jest"
   },
