+++ src/hooks/index.ts
@@ -1,0 +1,26 @@
+export { useFetcher as useFetch } from "./use-fetcher"
+
+export {
+  useFetcherLoading as useLoading,
+  useFetcherConfig as useConfig,
+  useFetcherData as useData,
+  useFetcherCode as useCode,
+  useFetcherError as useError,
+  useFetcherMutate as useMutate,
+  useFetcherId as useFetchId,
+  useFetcherBlob as useBlob,
+  useFetcherText as useText,
+  useGET,
+  useDELETE,
+  useHEAD,
+  useOPTIONS,
+  usePOST,
+  usePUT,
+  usePATCH,
+  usePURGE,
+  useLINK,
+  useUNLINK,
+  useResolve,
+  useGql,
+  useImperative,
+} from "./others"
